move pu weight producer
