+++ src/classes.h
@@ -1,7 +1,9 @@
 #include "PhysicsTools/NanoAODTools/interface/PyJetResolutionWrapper.h"
 #include "PhysicsTools/NanoAODTools/interface/PyJetResolutionScaleFactorWrapper.h"
 #include "PhysicsTools/NanoAODTools/interface/PyJetParametersWrapper.h"
+#include "PhysicsTools/NanoAODTools/interface/WeightCalculatorFromHistogram.h"
 
 PyJetResolutionWrapper jetRes;
 PyJetResolutionScaleFactorWrapper jetResScaleFactor;
 PyJetParametersWrapper jetParams;
+WeightCalculatorFromHistogram wcalc;
